Refactor users routes
