+++ src/app.ts
@@ -1,11 +1,11 @@
 import fastify from 'fastify'
-import { appRoutes } from './http/routes'
+import { usersRoutes } from './http/controllers/users/routes'
 import { ZodError } from 'zod'
 import { env } from './env'
 
 export const app = fastify()
 
-app.register(appRoutes)
+app.register(usersRoutes)
 
 app.setErrorHandler((error, _, res) => {
   if (error instanceof ZodError) {
