+++ src/http/controllers/users/routes.ts
@@ -1,0 +1,6 @@
+import { FastifyInstance } from 'fastify'
+import { register } from './register'
+
+export async function usersRoutes(app: FastifyInstance) {
+  app.post('/users', register)
+}
