--- src/http/routes.ts
@@ -1,6 +1,0 @@
-import { FastifyInstance } from 'fastify'
-import { register } from './controllers/register'
-
-export async function appRoutes(app: FastifyInstance) {
-  app.post('/users', register)
-}
